ajout des routes  association demande kanban et produit

+++ models/frspdt.js
@@ -1,21 +1,21 @@
 module.exports = (sequelize, DataTypes) => {
   const Frspdt = sequelize.define("frspdt", {
     id: {
-      type: DataTypes.INTEGER(10),
+      type: DataTypes.INTEGER,
       primaryKey: true,
       autoIncrement: true,
     },
 
     delai_jour: {
-      type: DataTypes.INTEGER(10),
+      type: DataTypes.INTEGER,
       allowNul: false,
     },
     quantite_min: {
-      type: DataTypes.INTEGER(10),
+      type: DataTypes.INTEGER,
       allowNul: false,
     },
     quantite_unite: {
-      type: DataTypes.INTEGER(10),
+      type: DataTypes.INTEGER,
       allowNul: false,
     },
   });
